fractional pad/column to give 1-1 correspondence

+++ StRoot/StRrsMaker/StRichRawCoordinate.h
@@ -1,6 +1,6 @@
 /***********************************************************************
  *
- * $Id: StRichRawCoordinate.h,v 1.1 2000/02/08 16:34:12 lasiuk Exp $
+ * $Id: StRichRawCoordinate.h,v 1.2 2000/04/05 16:03:58 lasiuk Exp $
  *
  * Author:  brian Jan 27, 2000
  *
@@ -11,8 +11,8 @@
  *************************************************************************
  *
  * $Log: StRichRawCoordinate.h,v $
- * Revision 1.1  2000/02/08 16:34:12  lasiuk
- * Initial Revision:  eventually for StUtilities
+ * Revision 1.2  2000/04/05 16:03:58  lasiuk
+ * fractional pad/column to give 1-1 correspondence
  *
  * Revision 1.1  2000/02/08 16:34:12  lasiuk
  * Initial Revision:  eventually for StUtilities
@@ -27,31 +27,31 @@
 class StRichRawCoordinate {
 public:
     StRichRawCoordinate();
-    StRichRawCoordinate(const int pad, const int row);
+    StRichRawCoordinate(const double pad, const double row);
 
     virtual ~StRichRawCoordinate();
     //StRichRawCoordinate(const StRichRawCoordinate&);
     //StRichRawCoordinate& operator=(cont StRichRawCoordinate&);
     
     // access functions
-    const int row()     const;
-    const int pad()     const;
-    const int column()  const;
-    void setRow(int);
-    void setPad(int);
-    void setColumn(int);
+    const double row()     const;
+    const double pad()     const;
+    const double column()  const;
+    void setRow(double);
+    void setPad(double);
+    void setColumn(double);
     
 private:
-    int mRow;
-    int mPad;
+    double mRow;
+    double mPad;
 };
 
-const inline int StRichRawCoordinate::row()        const {return(mRow);}
-const inline int StRichRawCoordinate::pad()        const {return(mPad);}
-const inline int StRichRawCoordinate::column()     const {return(mPad);}
-inline void StRichRawCoordinate::setRow(int r)    {mRow = r;}
-inline void StRichRawCoordinate::setPad(int p)    {mPad = p;}
-inline void StRichRawCoordinate::setColumn(int p) {mPad = p;}
+const inline double StRichRawCoordinate::row()        const {return(mRow);}
+const inline double StRichRawCoordinate::pad()        const {return(mPad);}
+const inline double StRichRawCoordinate::column()     const {return(mPad);}
+inline void StRichRawCoordinate::setRow(double r)    {mRow = r;}
+inline void StRichRawCoordinate::setPad(double p)    {mPad = p;}
+inline void StRichRawCoordinate::setColumn(double p) {mPad = p;}
 // Non-member
 ostream& operator<<(ostream&, const StRichRawCoordinate&);
